// frontend/lib/api/types.ts
export interface components {
    schemas: {
// ...
        Subscription: {
            /** Format: uuid */
            id?: string;
            /** Format: uuid */
            customer_id?: string;
            plan?: components["schemas"]["Plan"];
            /** @enum {string} */
            status?: "incomplete" | "active" | "past_due" | "canceled" | "unpaid";
            /** Format: date-time */
            current_period_start?: string | null;
            /** Format: date-time */
            current_period_end?: string | null;
            cancel_at_period_end?: boolean;
            /** Format: date-time */
            canceled_at?: string | null;
            /** Format: date-time */
            created_at?: string;
        };
// ...
    };
// ...
}